% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pypresence}
\alias{check_pypresence}
\title{Check if pypresence is Installed}
\usage{
check_pypresence(python_path)
}
\arguments{
\item{python_path}{Path to Python executable}
}
\value{
Logical indicating if pypresence is installed
}
\description{
Verifies that the pypresence Python library is installed
}
\keyword{internal}
