% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{configure_rprofile}
\alias{configure_rprofile}
\title{Configure .Rprofile for Auto-Start}
\usage{
configure_rprofile()
}
\value{
Logical indicating success
}
\description{
Adds startup code to the user's .Rprofile
}
\keyword{internal}
