% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_pypresence}
\alias{install_pypresence}
\title{Install pypresence Library}
\usage{
install_pypresence(python_path)
}
\arguments{
\item{python_path}{Path to Python executable}
}
\value{
Logical indicating success
}
\description{
Attempts to install the pypresence Python library
}
\keyword{internal}
