% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{Start Discord Rich Presence Integration}
\usage{
start()
}
\description{
Initializes the Discord Rich Presence integration for RStudio.
This function starts a periodic monitoring process that updates your Discord
status based on the currently active file in RStudio.
}
\examples{
\dontrun{
discordrpc::start()
}
}
