% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update Discord Rich Presence}
\usage{
update()
}
\description{
Checks the active editor file and updates Discord Rich Presence.
This function is called periodically by the later package and is not
intended to be called directly by the user.
}
\keyword{internal}
